Creating the first layer and prediction generation

+++ JAVA/Cognition/src/main/java/CognitionMap/Elements/Bias.java
@@ -2,7 +2,7 @@ package CognitionMap.Elements;
 
 public class Bias extends Cognition_Element {
 
-    public Bias(int value) {
+    public Bias(float value) {
         super(value);
     }
 }
