+++ JAVA/Cognition/src/main/java/CognitionMap/Elements/Weight.java
@@ -1,7 +1,7 @@
 package CognitionMap.Elements;
 
 public class Weight extends Cognition_Element {
-    public Weight(int value) {
+    public Weight(float value) {
         super(value);
     }
 }
